backend/requirements.txt
fastapi
psycopg2-binary
email-validator
python-dotenv
uvicorn
sqlalchemy
